packages/sprite-tools/requirements.txt
# Sprite extraction dependencies
# Install: pip install -r requirements.txt

# PyTorch (MPS for Mac, CUDA for Linux/Windows)
torch>=2.0.0
torchvision>=0.15.0

# BiRefNet (GPU-accelerated background removal, Apache 2.0)
transformers>=4.39.0
huggingface_hub>=0.20.0
einops>=0.8.0
kornia>=0.7.0
timm>=1.0.0

# Image processing
pillow>=10.0.0
numpy>=1.24.0

# Background removal (CPU, uses U2-Net) - RECOMMENDED for cartoon sprites
rembg>=2.0.50
onnxruntime>=1.16.0

# Loop detection (LoopyCut dependencies)
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
ffmpeg-python>=0.2.0
click>=8.1.0
numba>=0.58.0

# Video processing
# Note: ffmpeg must be installed via system package manager (brew install ffmpeg)
